LGTM画像作成処理を main.go から create_lgtm_image.go に移動
